+++ Localization/TellMeWhen-enUS.lua
@@ -631,11 +631,13 @@ L["RUNES"] = "Rune(s) to check"
 
 
 L["ICONMENU_SHOWTTTEXT2"] = "Aura variables"
-L["ICONMENU_SHOWTTTEXT_DESC2"] = [[Report the icon's stacks as a variable associated with the aura. Practical uses include monitoring damage shield amounts.
+L["ICONMENU_SHOWTTTEXT_DESC2"] = [[Choose which value will be reported as the icon's stacks.
 
-This value will be reported and shown in place of the icon's stack count.
-
-Numbers are provided by Blizzard API and do not necessarily match numbers found on the tooltip of the aura.]]
+Choose between:
+• The actual aura stacks
+• Variables reported by Blizzard's API
+• Numbers extracted from the aura's tooltip
+]]
 
 L["ICONMENU_SHOWTTTEXT_STACKS"] = "Stacks (default behavior)"
 L["ICONMENU_SHOWTTTEXT_STACKS_DESC"] = "Causes the buff/debuff's stacks to be reported as the icon's stacks."
@@ -643,10 +645,15 @@ L["ICONMENU_SHOWTTTEXT_FIRST"] = "First non-zero variable"
 L["ICONMENU_SHOWTTTEXT_FIRST_DESC"] = [[Causes the first non-zero variable associated with the buff/debuff to be reported as the icon's stacks.
 
 Usually this will be the correct variable if you desire one of the aura's variables.]]
-L["ICONMENU_SHOWTTTEXT_VAR"] = "Only Variable #%d"
+L["ICONMENU_SHOWTTTEXT_VAR"] = "Variable #%d"
 L["ICONMENU_SHOWTTTEXT_VAR_DESC"] = [[Causes only this variable to be reported as the icon's stacks.
 
 Use this if other, incorrect variables are sometimes reported. Use trial-and-error to figure out which of the variables is correct.]]
+
+L["ICONMENU_SHOWTTTEXT_TT"] = "Tooltip #%d"
+L["ICONMENU_SHOWTTTEXT_TT_DESC"] = [[Causes only the number at position %d in the aura's tooltip to be reported as the icon's stacks.
+
+Tooltip scanning consumes additional resources - only use when the Variable numbers aren't being provided by Blizzard.]]