+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v11.0.5
 * Spell Cooldown and Reactive Ability icons are vastly more efficient if the tracked ability is on your action bars (macros excluded).
   * This is done by utilizing new WoW 11.0 APIs that are specific to abilities present on action bars.
+* #2190: Added options to Buff/Debuff icons to source stack count from tooltip numbers.
 * Fix: #2197 upstream issue in LibSpellRange-1.0 with range checking in classic/cata.
 * Fix: #2201 Don't treat inactive hero talent trees as learned talents